+++ web/lib/src/dom/ext_color_buffer_half_float.dart
@@ -18,8 +18,11 @@ import 'dart:js_interop';
 import 'webgl1.dart';
 
 extension type EXT_color_buffer_half_float._(JSObject _) implements JSObject {
-  external static GLenum get RGBA16F_EXT;
-  external static GLenum get RGB16F_EXT;
-  external static GLenum get FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT;
-  external static GLenum get UNSIGNED_NORMALIZED_EXT;
+  static const GLenum RGBA16F_EXT = 34842;
+
+  static const GLenum RGB16F_EXT = 34843;
+
+  static const GLenum FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT = 33297;
+
+  static const GLenum UNSIGNED_NORMALIZED_EXT = 35863;
 }