+++ web/lib/src/dom/filter_effects.dart
@@ -56,23 +56,40 @@ extension type SVGFEBlendElement._(JSObject _) implements SVGElement, JSObject {
           'feBlend',
         );
 
-  external static int get SVG_FEBLEND_MODE_UNKNOWN;
-  external static int get SVG_FEBLEND_MODE_NORMAL;
-  external static int get SVG_FEBLEND_MODE_MULTIPLY;
-  external static int get SVG_FEBLEND_MODE_SCREEN;
-  external static int get SVG_FEBLEND_MODE_DARKEN;
-  external static int get SVG_FEBLEND_MODE_LIGHTEN;
-  external static int get SVG_FEBLEND_MODE_OVERLAY;
-  external static int get SVG_FEBLEND_MODE_COLOR_DODGE;
-  external static int get SVG_FEBLEND_MODE_COLOR_BURN;
-  external static int get SVG_FEBLEND_MODE_HARD_LIGHT;
-  external static int get SVG_FEBLEND_MODE_SOFT_LIGHT;
-  external static int get SVG_FEBLEND_MODE_DIFFERENCE;
-  external static int get SVG_FEBLEND_MODE_EXCLUSION;
-  external static int get SVG_FEBLEND_MODE_HUE;
-  external static int get SVG_FEBLEND_MODE_SATURATION;
-  external static int get SVG_FEBLEND_MODE_COLOR;
-  external static int get SVG_FEBLEND_MODE_LUMINOSITY;
+  static const int SVG_FEBLEND_MODE_UNKNOWN = 0;
+
+  static const int SVG_FEBLEND_MODE_NORMAL = 1;
+
+  static const int SVG_FEBLEND_MODE_MULTIPLY = 2;
+
+  static const int SVG_FEBLEND_MODE_SCREEN = 3;
+
+  static const int SVG_FEBLEND_MODE_DARKEN = 4;
+
+  static const int SVG_FEBLEND_MODE_LIGHTEN = 5;
+
+  static const int SVG_FEBLEND_MODE_OVERLAY = 6;
+
+  static const int SVG_FEBLEND_MODE_COLOR_DODGE = 7;
+
+  static const int SVG_FEBLEND_MODE_COLOR_BURN = 8;
+
+  static const int SVG_FEBLEND_MODE_HARD_LIGHT = 9;
+
+  static const int SVG_FEBLEND_MODE_SOFT_LIGHT = 10;
+
+  static const int SVG_FEBLEND_MODE_DIFFERENCE = 11;
+
+  static const int SVG_FEBLEND_MODE_EXCLUSION = 12;
+
+  static const int SVG_FEBLEND_MODE_HUE = 13;
+
+  static const int SVG_FEBLEND_MODE_SATURATION = 14;
+
+  static const int SVG_FEBLEND_MODE_COLOR = 15;
+
+  static const int SVG_FEBLEND_MODE_LUMINOSITY = 16;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedString get in2;
   external SVGAnimatedEnumeration get mode;
@@ -98,11 +115,16 @@ extension type SVGFEColorMatrixElement._(JSObject _)
           'feColorMatrix',
         );
 
-  external static int get SVG_FECOLORMATRIX_TYPE_UNKNOWN;
-  external static int get SVG_FECOLORMATRIX_TYPE_MATRIX;
-  external static int get SVG_FECOLORMATRIX_TYPE_SATURATE;
-  external static int get SVG_FECOLORMATRIX_TYPE_HUEROTATE;
-  external static int get SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA;
+  static const int SVG_FECOLORMATRIX_TYPE_UNKNOWN = 0;
+
+  static const int SVG_FECOLORMATRIX_TYPE_MATRIX = 1;
+
+  static const int SVG_FECOLORMATRIX_TYPE_SATURATE = 2;
+
+  static const int SVG_FECOLORMATRIX_TYPE_HUEROTATE = 3;
+
+  static const int SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA = 4;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedEnumeration get type;
   external SVGAnimatedNumberList get values;
@@ -147,12 +169,18 @@ extension type SVGFEComponentTransferElement._(JSObject _)
 /// [MDN Web Docs](https://developer.mozilla.org/en-US/docs/Web/API/SVGComponentTransferFunctionElement).
 extension type SVGComponentTransferFunctionElement._(JSObject _)
     implements SVGElement, JSObject {
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_UNKNOWN;
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_IDENTITY;
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_TABLE;
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_DISCRETE;
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_LINEAR;
-  external static int get SVG_FECOMPONENTTRANSFER_TYPE_GAMMA;
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_UNKNOWN = 0;
+
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_IDENTITY = 1;
+
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_TABLE = 2;
+
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_DISCRETE = 3;
+
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_LINEAR = 4;
+
+  static const int SVG_FECOMPONENTTRANSFER_TYPE_GAMMA = 5;
+
   external SVGAnimatedEnumeration get type;
   external SVGAnimatedNumberList get tableValues;
   external SVGAnimatedNumber get slope;
@@ -241,13 +269,20 @@ extension type SVGFECompositeElement._(JSObject _)
           'feComposite',
         );
 
-  external static int get SVG_FECOMPOSITE_OPERATOR_UNKNOWN;
-  external static int get SVG_FECOMPOSITE_OPERATOR_OVER;
-  external static int get SVG_FECOMPOSITE_OPERATOR_IN;
-  external static int get SVG_FECOMPOSITE_OPERATOR_OUT;
-  external static int get SVG_FECOMPOSITE_OPERATOR_ATOP;
-  external static int get SVG_FECOMPOSITE_OPERATOR_XOR;
-  external static int get SVG_FECOMPOSITE_OPERATOR_ARITHMETIC;
+  static const int SVG_FECOMPOSITE_OPERATOR_UNKNOWN = 0;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_OVER = 1;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_IN = 2;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_OUT = 3;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_ATOP = 4;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_XOR = 5;
+
+  static const int SVG_FECOMPOSITE_OPERATOR_ARITHMETIC = 6;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedString get in2;
   external SVGAnimatedEnumeration get operator;
@@ -277,10 +312,14 @@ extension type SVGFEConvolveMatrixElement._(JSObject _)
           'feConvolveMatrix',
         );
 
-  external static int get SVG_EDGEMODE_UNKNOWN;
-  external static int get SVG_EDGEMODE_DUPLICATE;
-  external static int get SVG_EDGEMODE_WRAP;
-  external static int get SVG_EDGEMODE_NONE;
+  static const int SVG_EDGEMODE_UNKNOWN = 0;
+
+  static const int SVG_EDGEMODE_DUPLICATE = 1;
+
+  static const int SVG_EDGEMODE_WRAP = 2;
+
+  static const int SVG_EDGEMODE_NONE = 3;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedInteger get orderX;
   external SVGAnimatedInteger get orderY;
@@ -408,11 +447,16 @@ extension type SVGFEDisplacementMapElement._(JSObject _)
           'feDisplacementMap',
         );
 
-  external static int get SVG_CHANNEL_UNKNOWN;
-  external static int get SVG_CHANNEL_R;
-  external static int get SVG_CHANNEL_G;
-  external static int get SVG_CHANNEL_B;
-  external static int get SVG_CHANNEL_A;
+  static const int SVG_CHANNEL_UNKNOWN = 0;
+
+  static const int SVG_CHANNEL_R = 1;
+
+  static const int SVG_CHANNEL_G = 2;
+
+  static const int SVG_CHANNEL_B = 3;
+
+  static const int SVG_CHANNEL_A = 4;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedString get in2;
   external SVGAnimatedNumber get scale;
@@ -492,10 +536,14 @@ extension type SVGFEGaussianBlurElement._(JSObject _)
           'feGaussianBlur',
         );
 
-  external static int get SVG_EDGEMODE_UNKNOWN;
-  external static int get SVG_EDGEMODE_DUPLICATE;
-  external static int get SVG_EDGEMODE_WRAP;
-  external static int get SVG_EDGEMODE_NONE;
+  static const int SVG_EDGEMODE_UNKNOWN = 0;
+
+  static const int SVG_EDGEMODE_DUPLICATE = 1;
+
+  static const int SVG_EDGEMODE_WRAP = 2;
+
+  static const int SVG_EDGEMODE_NONE = 3;
+
   external void setStdDeviation(
     num stdDeviationX,
     num stdDeviationY,
@@ -589,9 +637,12 @@ extension type SVGFEMorphologyElement._(JSObject _)
           'feMorphology',
         );
 
-  external static int get SVG_MORPHOLOGY_OPERATOR_UNKNOWN;
-  external static int get SVG_MORPHOLOGY_OPERATOR_ERODE;
-  external static int get SVG_MORPHOLOGY_OPERATOR_DILATE;
+  static const int SVG_MORPHOLOGY_OPERATOR_UNKNOWN = 0;
+
+  static const int SVG_MORPHOLOGY_OPERATOR_ERODE = 1;
+
+  static const int SVG_MORPHOLOGY_OPERATOR_DILATE = 2;
+
   external SVGAnimatedString get in1;
   external SVGAnimatedEnumeration get operator;
   external SVGAnimatedNumber get radiusX;
@@ -695,12 +746,18 @@ extension type SVGFETurbulenceElement._(JSObject _)
           'feTurbulence',
         );
 
-  external static int get SVG_TURBULENCE_TYPE_UNKNOWN;
-  external static int get SVG_TURBULENCE_TYPE_FRACTALNOISE;
-  external static int get SVG_TURBULENCE_TYPE_TURBULENCE;
-  external static int get SVG_STITCHTYPE_UNKNOWN;
-  external static int get SVG_STITCHTYPE_STITCH;
-  external static int get SVG_STITCHTYPE_NOSTITCH;
+  static const int SVG_TURBULENCE_TYPE_UNKNOWN = 0;
+
+  static const int SVG_TURBULENCE_TYPE_FRACTALNOISE = 1;
+
+  static const int SVG_TURBULENCE_TYPE_TURBULENCE = 2;
+
+  static const int SVG_STITCHTYPE_UNKNOWN = 0;
+
+  static const int SVG_STITCHTYPE_STITCH = 1;
+
+  static const int SVG_STITCHTYPE_NOSTITCH = 2;
+
   external SVGAnimatedNumber get baseFrequencyX;
   external SVGAnimatedNumber get baseFrequencyY;
   external SVGAnimatedInteger get numOctaves;