+++ web/lib/src/dom/webidl.dart
@@ -42,31 +42,55 @@ extension type DOMException._(JSObject _) implements JSObject {
     String name,
   ]);
 
-  external static int get INDEX_SIZE_ERR;
-  external static int get DOMSTRING_SIZE_ERR;
-  external static int get HIERARCHY_REQUEST_ERR;
-  external static int get WRONG_DOCUMENT_ERR;
-  external static int get INVALID_CHARACTER_ERR;
-  external static int get NO_DATA_ALLOWED_ERR;
-  external static int get NO_MODIFICATION_ALLOWED_ERR;
-  external static int get NOT_FOUND_ERR;
-  external static int get NOT_SUPPORTED_ERR;
-  external static int get INUSE_ATTRIBUTE_ERR;
-  external static int get INVALID_STATE_ERR;
-  external static int get SYNTAX_ERR;
-  external static int get INVALID_MODIFICATION_ERR;
-  external static int get NAMESPACE_ERR;
-  external static int get INVALID_ACCESS_ERR;
-  external static int get VALIDATION_ERR;
-  external static int get TYPE_MISMATCH_ERR;
-  external static int get SECURITY_ERR;
-  external static int get NETWORK_ERR;
-  external static int get ABORT_ERR;
-  external static int get URL_MISMATCH_ERR;
-  external static int get QUOTA_EXCEEDED_ERR;
-  external static int get TIMEOUT_ERR;
-  external static int get INVALID_NODE_TYPE_ERR;
-  external static int get DATA_CLONE_ERR;
+  static const int INDEX_SIZE_ERR = 1;
+
+  static const int DOMSTRING_SIZE_ERR = 2;
+
+  static const int HIERARCHY_REQUEST_ERR = 3;
+
+  static const int WRONG_DOCUMENT_ERR = 4;
+
+  static const int INVALID_CHARACTER_ERR = 5;
+
+  static const int NO_DATA_ALLOWED_ERR = 6;
+
+  static const int NO_MODIFICATION_ALLOWED_ERR = 7;
+
+  static const int NOT_FOUND_ERR = 8;
+
+  static const int NOT_SUPPORTED_ERR = 9;
+
+  static const int INUSE_ATTRIBUTE_ERR = 10;
+
+  static const int INVALID_STATE_ERR = 11;
+
+  static const int SYNTAX_ERR = 12;
+
+  static const int INVALID_MODIFICATION_ERR = 13;
+
+  static const int NAMESPACE_ERR = 14;
+
+  static const int INVALID_ACCESS_ERR = 15;
+
+  static const int VALIDATION_ERR = 16;
+
+  static const int TYPE_MISMATCH_ERR = 17;
+
+  static const int SECURITY_ERR = 18;
+
+  static const int NETWORK_ERR = 19;
+
+  static const int ABORT_ERR = 20;
+
+  static const int URL_MISMATCH_ERR = 21;
+
+  static const int QUOTA_EXCEEDED_ERR = 22;
+
+  static const int TIMEOUT_ERR = 23;
+
+  static const int INVALID_NODE_TYPE_ERR = 24;
+
+  static const int DATA_CLONE_ERR = 25;
 
   /// The **`name`** read-only property of the
   /// [DOMException] interface returns a string that contains